# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Affiliate Product Discovery Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSpecifications } from './productSpecifications';

export interface Product {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  currency?: string;
  images?: string[];
  /** @nullable */
  brand?: string | null;
  category: string;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  affiliateUrl: string;
  /** @nullable */
  store?: string | null;
  /** @nullable */
  specifications?: ProductSpecifications;
  clickCount: number;
  featured: boolean;
  inStock?: boolean;
  /** @nullable */
  sourceUrl?: string | null;
  createdAt: string;
}