# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Affiliate Product Discovery Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputSpecifications } from './productInputSpecifications';

export interface ProductInput {
  title: string;
  description?: string;
  price: number;
  originalPrice?: number;
  currency?: string;
  images?: string[];
  brand?: string;
  category: string;
  rating?: number;
  reviewCount?: number;
  affiliateUrl: string;
  store?: string;
  specifications?: ProductInputSpecifications;
  featured?: boolean;
  inStock?: boolean;
  sourceUrl?: string;
}